restructured waqar components
