+++ src/components/error/error404.jsx
